+++ jdi-light-vuetify-tests/src/main/java/io/github/com/pages/SheetsPage.java
@@ -1,7 +1,7 @@
 package io.github.com.pages;
 
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
-import com.epam.jdi.light.vuetify.elements.composite.Sheet;
+import com.epam.jdi.light.vuetify.elements.common.Sheet;
 
 import java.util.List;
